Post conversation on jan 30, i am updaating HIP-RA so that you can see a version that doesn't calculate negative enthalpies but may still be wrong.
This update to HIP-RA-X adds the input variables to the report, as we discussed during our last meeting. I also added two more calculations and reported values: heat per unit volme and power per unit volume for the reservoir. This is essentially an energy density calculation. This passes the tests (after I updated the .out file to look right with the new reported lines).